backend: Numeric/Decimal para todos os valores monetarios (D-018)
- cost_lines.amount e material_request_items.unit_price: Float -> Numeric(12, 2)
  (Decimal em Python). Evita erro de arredondamento binario
  (0.1 + 0.1 + 0.1 != 0.3 em float) em valores de dinheiro.
- FinancialCostRecord.amount segue o mesmo tipo; CsvFinancialAdapter le o
  CSV com Decimal(str) diretamente, nunca float(...) a meio do caminho.
- Quantidades (inventory_movements.quantity, material_request_items.quantity,
  inventory_items.min_stock) mantidas em Float de proposito — nao sao
  valores monetarios.
- Testes: caso que demonstra o erro real que float cometeria, round-trip
  em SQLite como Decimal exato, e soma exata a partir do CSV sintetico.

+++ backend/app/adapters/financial/base.py
@@ -8,6 +8,7 @@ export estruturado), API/Excel ficam como interface preparada.
 from __future__ import annotations
 
 from dataclasses import dataclass
+from decimal import Decimal
 from typing import Protocol
 
 
@@ -15,7 +16,9 @@ from typing import Protocol
 class FinancialCostRecord:
     external_project_ref: str
     category: str
-    amount: float
+    # Decimal, nunca float — ver app/models/cost.py:MONEY e
+    # tests/test_monetary_precision.py.
+    amount: Decimal
     currency: str
     reference: str | None = None
